Kjører utledning av fagsaker for overgang til høy sats i egen task. Utvider med mulighet for å filtrere for å kjøre for enkelt fagsak.
